Faq page navigation improvement
Is difficoult scroll a ton of content on a mobile device screen, so I
added a navigation system on the faq page.

It's composed by an index and the classic top button to go back on top
of the page.

Couse a big question needs a big answer, I separate some wall of text in
2 paragraph for the sake of readability.

It's not a big deal on desktop, but on mobile it is.

I added the necessary CSS code to style a little bit.

NOTE: some links were ponting to the old repository website, I've update
but still don't know what use to substitute the forum links. Let me
know.