Images arrangement and classification
It's better to have all images needed in one folder "images"

I've kept the original subfolders "logos" and "screenshot" but now in
the image folder.

Also GitHub logo was moved into "logos" folder with its similar, be
alone is not good!

That wasn't a very important upgrade, but helpful in work to come.