Added Meta tag "description" to all html pages
Simply added the Meta tag "description" to the code. That allow us to
choose (or hopefully choose) the content of the text visible on the SERP
under the title.

If is not defined from the code, the search engine will abtract an
arbitrary part of text. Not always the best.