Added Header/Content/Footer page template
I simply rearranged the html code inside the pages in the traditional
form:
<body>
<header></header>
<div id=content></div>
<footer></footer>
</body>

Some CSS rearrangement was necessary, but nothing revolutionary.

The rendering look of the page is identical except concerning the footer
(license). I like this calssic template more, but it is possibile and
very easy apply the old and include license in the white background
zone. Just let me know.