Multilanguage-tool and index.html italian translation (test)
Created a multilanguage_tool directory, hide from the bots, that contain
a php version of the website where content are aside from the template.

In the languages directory, each part of the page (index.html is just
the first) is translated.

Common parts (header and footer) will be translated once, and a php file
named like the page will contain the content definition about that page
in the requested language.